% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_citation.R
\name{fmt_citation}
\alias{fmt_citation}
\title{Accepts a bibtex file, returns a formatted citation/bibtex key join table, meant to supplement the gt format functions}
\usage{
fmt_citation(bibtex_in)
}
\arguments{
\item{bibtex_in}{bibtex file to read in}
}
\value{
join table between BIBTEXKEY and citation
}
\description{
Accepts a bibtex file, returns a formatted citation/bibtex key join table, meant to supplement the gt format functions
}
\examples{
software.yaml <- yaml::read_yaml("data/softwareUsed.yaml")
software.df <- plyr::ldply(software.yaml$software, data.frame)
join.tbl <- fmt_citation("data/software.bib")
inner_join(software.df \%>\% separate_rows(bibtex) , join.tbl, by=c("bibtex"="BIBTEXKEY")) \%>\% group_by(name) \%>\%  mutate(allCit = paste0(cite, collapse = "; ")) \%>\% select(-c(bibtex,cite)) \%>\% unique() \%>\% ungroup() \%>\%  gt::gt() \%>\% tab_header(title="Bioinformatics Software Used", subtitle= md("&nbsp;") )
}
