% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_readable_croncher.R, R/save_table.R
\name{human_readable_croncher}
\alias{human_readable_croncher}
\title{Reformat very large/small numbers as human-readable}
\usage{
human_readable_croncher(num_in)

human_readable_croncher(num_in)
}
\arguments{
\item{num_in}{A number.}
}
\value{
the number as string, human friendly

the number as string, human friendly
}
\description{
Reformat very large/small numbers as human-readable

Reformat very large/small numbers as human-readable
}
\examples{
human_readable_croncher(100000000)
human_readable_croncher(1/100000000.)
human_readable_croncher(100000000)
human_readable_croncher(1/100000000.)
}
