% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_figure.R
\name{save_figure}
\alias{save_figure}
\title{Wrapper to save a .PNG}
\usage{
save_figure(img_obj, fn)
}
\arguments{
\item{img_obj}{image to be saved}

\item{fn}{file to write}
}
\value{
nothing
}
\description{
Wrapper to save a .PNG
}
\examples{
plot.gg <- ggplot(pressure) + geom_point(aes(x=pressure, y=temperature))
save_figure(img_obj=plot.gg, fn="pressurePlot.png")
}
